Add a robust bluetooth sensor support

+++ app/src/main/java/studies/movie_pulse_app/sensor/event/ValueReading.java
@@ -5,9 +5,9 @@ package studies.movie_pulse_app.sensor.event;
  */
 public class ValueReading {
     public long time; // System time in milliseconds
-    public short value; // Pulse in 0-255
+    public byte value; // Pulse in 0-255
 
-    public ValueReading(long time, short value) {
+    public ValueReading(long time, byte value) {
         this.time = time;
         this.value = value;
     }
